Better way to handle DMs
Before, a DM would show the channel name as `undefined` because that is not a thing in the context of private channels.

This way, we assume its a direct message unless there's a guild, where we throw in the channel name and guild name in one var. It takes out the guildname var and the other var and just keeps it in one.